роуты, частичный mvc, вертска..-_-
