+++ project-1/src/App.css
@@ -1,4 +1,23 @@
+.App-logo {
+  height: 8vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
 .App {
+  background-color: rgba(40, 44, 52, 0.27);
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  /*justify-content: center;*/
+  font-size: calc(10px + 2vmin);
+  color: black;
   text-align: center;
 }
 
@@ -12,28 +31,6 @@
   /*justify-content: center;*/
   font-size: calc(10px + 2vmin);
   color: white;
-}
-
-.App-logo {
-  height: 8vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: rgba(40, 44, 52, 0.27);
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  /*justify-content: center;*/
-  font-size: calc(10px + 2vmin);
-  color: black;
 }
 
 .App-link {
