+++ project-1/src/components/footer/Footer.css
@@ -1,0 +1,13 @@
+.footer {
+    width: 100%;
+   /* padding: 1rem;*/
+    background-color: #FFFFFF;
+    border-radius: 0.1rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.footer_dark{
+    background-color: #383838;
+}
