+++ project-1/src/components/users/Users.css
@@ -1,14 +1,7 @@
 .users {
     display: flex;
     flex-wrap: wrap;
-    justify-content: space-between;
-}
-
-
-@media (min-width: 992px) {
-    .users{
-        justify-content: flex-start;
-    }
+    justify-content: center;
 }
 
 .container__item
@@ -20,8 +13,8 @@
     margin: 5px;
     padding: 0 0 5px 10px;
     box-shadow: 3px 3px 3px black;
-    width: 200px;
-    min-height: 200px;
+    width: 100px;
+    min-height: 100px;
 }
 
 .container__title
@@ -29,17 +22,6 @@
     display: flex;
     flex-direction: row;
     justify-content: space-between;
-}
-
-.container__delete
-{
-    background: none;
-    border-style: none;
-}
-
-.container__delete img
-{
-    height: 4vmin;
 }
 
 .container__item__text
